+++ lib/widgets/rixa_player.dart
@@ -1,22 +1,54 @@
+// ignore_for_file: unused_import
+
 import 'package:flutter/material.dart';
 import 'package:rixa/rixa.dart';
 import 'video_player_desktop.dart';
-import 'video_player_mobile.dart';
+//import 'video_player_mobile.dart';
 
 class RixaPlayer extends StatelessWidget {
-  final String path;
+  final String? filePath, url, asset;
   final double? width, height;
-  const RixaPlayer({
+  const RixaPlayer.file({
     Key? key,
-    required this.path,
+    required this.filePath,
     this.width,
     this.height,
-  }) : super(key: key);
+  })  : url = null,
+        asset = null,
+        super(key: key);
+  const RixaPlayer.network({
+    Key? key,
+    required this.url,
+    this.width,
+    this.height,
+  })  : filePath = null,
+        asset = null,
+        super(key: key);
+  const RixaPlayer.asset({
+    Key? key,
+    required this.asset,
+    this.width,
+    this.height,
+  })  : filePath = null,
+        url = null,
+        super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return Rixa.properties.isDesktop
-        ? VideoPlayerDesktop(path: path, width: width, height: height)
-        : VideoPlayerMobile(path: path, width: width, height: height);
+        ? VideoPlayerDesktop(
+            filePath: filePath,
+            url: url,
+            asset: asset,
+            width: width,
+            height: height)
+        : SizedBox(width: width, height: height);
+
+    //VideoPlayerMobile(
+    //    filePath: filePath,
+    //    url: url,
+    //    asset: asset,
+    //    width: width,
+    //    height: height);
   }
 }